requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
databases==0.5.5
ecdsa==0.17.0
fastapi==0.75.1
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
vk-api==11.9.7
watchgod==0.8
websockets==10.2
yarl==1.7.2
